minor bug fix

+++ config/INPD/features_transform.py
@@ -1,4 +1,5 @@
 import numpy as np
+import tensorflow as tf
 
 def features_transform(example):
     """ Receive and transform the features from features_dictionary.json"""
@@ -14,6 +15,6 @@ def features_transform(example):
     example['info/age_norm'] = example['info/age'] / 14.335387
 
     # create cl_tot normalized
-    example['info/cl_tot_norm'] = np.float32(example['info/cl_tot']) / 151.0
+    example['info/cl_tot_norm'] = tf.cast(example['info/cl_tot'], tf.float32) / 151.0
 
     return example
